+++ Manager.Domain/Commands/VendoresCommands/VendorListCommand.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Role { get; set; }
-        public double CustomCommission { get; set; }
+        public decimal CustomCommission { get; set; }
     }
 }
