+++ Manager.Domain/Entities/Vendor.cs
@@ -7,21 +7,21 @@ namespace Manager.Domain.Entities
         public string Name { get; private set; }
         public int RoleId { get; private set; }
         public Role Role { get; private set; }
-        public double CustomCommission { get; private set; }
+        public decimal CustomCommission { get; private set; }
 
         protected Vendor() {}
 
-        public Vendor(string name, Role role, double customCommission)
+        public Vendor(string name, Role role, decimal customCommission)
         {
             SetProperties(name, role, customCommission);
         }
 
-        public void Update(string name, Role role, double customCommsion)
+        public void Update(string name, Role role, decimal customCommsion)
         {
             SetProperties(name, role, customCommsion);
         }
 
-        private void SetProperties(string name, Role role, double customCommsion)
+        private void SetProperties(string name, Role role, decimal customCommsion)
         {
             Name = name;
             Role = role;
